+++ Realization/frontend/czechidm-core/src/content/role/Role.js
@@ -58,8 +58,8 @@ class Role extends Basic.AbstractContent {
     return (
       <div>
         <Basic.PageHeader showLoading={!entity && showLoading}>
-        <Basic.Icon value="fa:universal-access"/>
-        {' '}
+          <Basic.Icon value="fa:universal-access"/>
+          {' '}
           { manager.getNiceLabel(entity)} <small> {this.i18n('content.roles.edit.header') }</small>
         </Basic.PageHeader>
         <Basic.Row rendered={_requestsEnabled